pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-coder"
version = "0.1.0"
description = "CrewAI based Auto Coder SDK"
requires-python = ">=3.9"
dependencies = [
    "crewai",
    "langchain",
    "python-dotenv",
    "click",
    "pytest"
]

[project.scripts]
auto-coder = "auto_coder.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
